move from float to double

+++ src/idp/Filter.java
@@ -6,14 +6,14 @@ import java.awt.*;
  * Created by Andre on 23/04/2016.
  */
 public class Filter {
-    float scale;
+    double scale;
     public Filter(float scale) {
         this.scale = scale;
     }
-    float Frames(Frame f) {
+    double Frames(Frame f) {
         return 0;
     }
-    Color getColor(float f) {
+    Color getColor(double f) {
         return Color.black;
     }
 }
